Reorganize metadata: move scripts to _metadata/scripts/, move satellite_latencies.npy to _metadata/leo/, add README, remove obsolete scripts
